REFAC: rename utils/tools.py to toolbox.py to avoid confusion with ai tools
