+++ app/depends/orm_cache.py
@@ -16,7 +16,7 @@ from app.container import Get
 from app.services.database.redis_service import RedisService
 from app.utils.constant import RedisConstant
 from app.utils.helper import KeyBuilder
-from app.utils.tools import Time
+from app.utils.toolbox import Time
 from app.models.security_model import ClientORM,ChallengeORM, PolicyORM
 import typing
 from typing import Any, Callable, Generic, Type,TypeVar, TypedDict
