+++ app/ressources/agentic/data_ingest_ressource.py
@@ -34,7 +34,7 @@ from app.models.ingest_model import ResearchIngestUriMetadata,WebCrawlingIngestD
 from app.models.file_model import  FileResponseUploadModel, UploadError
 from app.data_ingest_tasks import DATA_TASK_REGISTRY_NAME
 from app.utils.constant import AgenticConstant, ArqDataTaskConstant, CostConstant
-from app.utils.tools import RunInThreadPool
+from app.utils.toolbox import RunInThreadPool
 from app.depends.variables import DeleteMode, force_update_query, delete_mode_query
