+++ app/ressources/object_s3_ressource.py
@@ -32,7 +32,7 @@ from app.utils.constant import SECONDS_IN_AN_HOUR as HOUR, CostConstant, MinioCo
 from app.utils.fileIO import ExtensionNotAllowedError, MultipleExtensionError
 from app.depends.variables import force_update_query
 from app.utils.helper import b64_encode
-from app.utils.tools import RunInThreadPool
+from app.utils.toolbox import RunInThreadPool
 
 @HTTPRessource('webhooks')
 class S3ObjectWebhookRessource(BaseHTTPRessource):
