+++ app/services/file/file_service.py
@@ -11,7 +11,7 @@ import base64
 import os
 import hashlib
 import shutil
-from app.utils.tools import RunInThreadPool
+from app.utils.toolbox import RunInThreadPool
 from app.utils.helper import b64_encode
 
 BytesSize= Literal['kb','mb','b']
