+++ app/services/worker/celery_service.py
@@ -19,7 +19,7 @@ from app.utils.constant import CeleryConstant, RedisConstant, SpecialKeyParamete
 import datetime as dt
 from humanize import naturaldelta,naturaltime
 from uuid import uuid4
-from app.utils.tools import RunInThreadPool
+from app.utils.toolbox import RunInThreadPool
 from app.classes.scheduler import schedule
 
 CHANNEL_BUILD_STATE=0
