+++ app/services/worker/task_service.py
@@ -15,7 +15,7 @@ from app.services.database.redis_service import RedisService
 from app.services.logger_service import LoggerService
 from app.services.vault_service import VaultService
 from app.utils.constant import MongooseDBConstant, RedisConstant, CeleryConstant
-from app.utils.tools import RunInThreadPool
+from app.utils.toolbox import RunInThreadPool
 from apscheduler.jobstores.base import JobLookupError
 from apscheduler.schedulers import (
     SchedulerAlreadyRunningError,
